be/pyproject.toml
[tool.poetry]
name = "be"
version = "0.1.0"
description = ""
authors = ["curekoshimizu"]

[tool.poetry.dependencies]
python = ">=3.8,<4"
fastapi = "^0.79.0"
uvicorn = "^0.18.2"
click = "^8.1.3"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
mypy = "^0.950"
pytest = "^6.2.1"
ipdb = "^0.13.4"
types-PyYAML = "^6.0.11"

[tool.black]
line-length = 120
# target-version = ["py38"]

[tool.isort]
default_section = "THIRDPARTY"
force_grid_wrap = 0
force_single_line = false
include_trailing_comma = true
known_first_party = ["be"]
line_length = 120
multi_line_output = 3
use_parentheses = true
